extract help text from commands.go
commands.go is very long making it hard to read and navigate. Break it
up by extracting help text and functions to its own file.
